+++ data.php
@@ -1,0 +1,42 @@
+<?php
+// показывать или нет выполненные задачи
+$show_complete_tasks = rand(0, 1);
+$projects = ["Входящие","Учеба","Работа","Домашние дела","Авто"];
+$tasks = [
+[
+	"title" => "Собеседование в IT компании",
+	"date" => "01.12.2018",
+	"project" => "Работа",
+	"status" => false
+],
+[
+	"title" => "Выполнить тестовое задание",
+	"date" => "25.12.2018",
+	"project" => "Работа",
+	"status" => false
+],
+[
+	"title" => "Сделать задание первого раздела",
+	"date" => "21.12.2018",
+	"project" => "Учеба",
+	"status" => true
+],
+[
+	"title" => "Встреча с другом",
+	"date" => "22.12.2018",
+	"project" => "Входящие",
+	"status" => false
+],
+[
+	"title" => "Купить корм для кота",
+	"date" => "Нет",
+	"project" => "Домашние дела",
+	"status" => false
+],
+[
+	"title" => "Заказать пиццу",
+	"date" => "Нет",
+	"project" => "Домашние дела",
+	"status" => false
+]
+];
