+++ templates/index.php
@@ -27,7 +27,7 @@
                         <td class="task__select">
                             <label class="checkbox task__checkbox">
                                 <input class="checkbox__input visually-hidden task__checkbox" type="checkbox" value="1" <?php if($task["status"]):?>checked<?php endif ?>>
-                                <span class="checkbox__text"><?=esc($task["title"])?></span>
+                                <span class="checkbox__text"><?=htmlspecialchars($task["title"])?></span>
                             </label>
                         </td>
                         <td class="task__date"><?=$task["date"]?></td>
