+++ templates/layout.php
@@ -24,7 +24,7 @@
 
                 <div class="main-header__side-item user-menu">
                     <div class="user-menu__data">
-                        <p><?=$user_name;?></p>
+                        <p><?=htmlspecialchars($user_name);?></p>
 
                         <a href="#">Выйти</a>
                     </div>
@@ -40,7 +40,7 @@
                     <ul class="main-navigation__list">
 						<?php foreach ($projects as $project):?>
                         <li class="main-navigation__list-item">
-                            <a class="main-navigation__list-item-link" href="#"><?=esc($project)?></a>
+                            <a class="main-navigation__list-item-link" href="#"><?=htmlspecialchars($project)?></a>
                             <span class="main-navigation__list-item-count"><?=count_tasks($tasks,$project)?></span>
                         </li>
 						<?php endforeach ?>
